requirements.txt
Flask==1.1.2
gunicorn==20.0.4
boto==2.49.0
discord.py== 1.3.3
dnspython==1.16.0
python-dotenv==0.13.0
PyNaCl==1.3.0
requests==2.23.0
async-timeout==3.0.1
unidecode==1.1.1
DateTime~=4.3
aiohttp~=3.6.2
youtube-dl~=2020.5.3